refactor: 타입 선언 위치 변경에 따른 import 수정
