+++ src/pages/SchedulePage/CalendarSection/CalenderConents.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DailyScheduleData } from 'apis/schedule/getMonthly';
+import { DailyWorkTimeData } from 'apis/types';
 import { MonthBox, WeekGrid } from 'components/Calendar/CalendarStyle';
 import CalendarDayBox from './CalendarDayBox';
 import useSchedule from 'hooks/SchedulePage/useSchedule';
@@ -8,9 +8,9 @@ const CalenderConents = (): JSX.Element => {
   const { scheduleData, selectedDate, dateOnClick } = useSchedule();
   return (
     <MonthBox $wFull>
-      {scheduleData?.table.map((weekArray: DailyScheduleData[], i) => (
+      {scheduleData?.table.map((weekArray: DailyWorkTimeData[], i) => (
         <WeekGrid key={`${i}주`}>
-          {weekArray.map((e: DailyScheduleData) => (
+          {weekArray.map((e: DailyWorkTimeData) => (
             <CalendarDayBox
               key={e.date}
               dateString={e.date}
