+++ src/pages/alba/ApplyPage/index.tsx
@@ -2,13 +2,12 @@ import React, { Suspense } from 'react';
 import { useLocation } from 'react-router-dom';
 import PageContainer from 'components/@commons/PageContainer';
 import { atom, useAtomValue } from 'jotai';
-import { SelectedSchedule } from 'apis/alba/apply';
 import weekdayArray from 'utils/weekdayArray';
 import TimeSelectSection from './TimeSelectSection';
-import { TimeData } from 'apis/admin/application/open';
 import PreviewSection from './PreviewSection';
+import { SelectedTimeData, TimeData } from 'apis/types';
 
-export const weeklySelectAtom = atom<SelectedSchedule[][]>(weekdayArray.map(() => []));
+export const weeklySelectAtom = atom<SelectedTimeData[][]>(weekdayArray.map(() => []));
 export const timeTemplateAtom = atom<{ [index: number]: TimeData }>({});
 export const applyStepAtom = atom(1);
